Finish mission6
