+++ README.md
@@ -8,44 +8,42 @@
 
 ### 체크리스트 [기본]
 
-#### 중고마켓
+#### 상품 등록
 
-- [] 중고마켓 페이지 주소는 “/items” 입니다.
-- [] 페이지 주소가 “/items” 일때 상단네비게이션바의 '중고마켓' 버튼의 색상은 “3692FF”입니다.
-- [x] 상단 네비게이션 바는 이전 미션에서 구현한 랜딩 페이지와 동일한 스타일로 만들어 주세요.
-- [x] 상품 데이터 정보는 https://panda-market-api.vercel.app/docs/#/ 에 명세된 GET 메소드 “/products” 를 사용해주세요.
-- [] '상품 등록하기' 버튼을 누르면 “/additem” 로 이동합니다. ( 빈 페이지 )
-- [x] 전체 상품에서 드롭 다운으로 “최신 순” 또는 “좋아요 순”을 선택해서 정렬을 할 수 있습니다.
-
-#### 중고마켓 반응형
-
-- [x] 베스트 상품
-- [x] 전체 상품
+- [x] 상품 등록 페이지 주소는 “/additem” 입니다.
+- [x] 페이지 주소가 “/additem” 일때 상단네비게이션바의 '중고마켓' 버튼의 색상은 “3692FF”입니다.
+- [x] 상품 이미지는 최대 한개 업로드가 가능합니다.
+- [x] 각 input의 placeholder 값을 정확히 입력해주세요.
+- [x] 이미지를 제외하고 input 에 모든 값을 입력하면 ‘등록' 버튼이 활성화 됩니다.
+- [x] API를 통한 상품 등록은 추후 미션에서 적용합니다.
 
 ### 체크리스트 [심화]
 
-- [x] 페이지 네이션 기능을 구현합니다.
+- [x] 이미지 안의 X 버튼을 누르면 이미지가 삭제됩니다.
+- [x] 추가된 태그 안의 X 버튼을 누르면 해당 태그는 삭제됩니다.
 
-### 멘토님께 
-- 페이지 이동 부분은 관련 개념을 더 학습하고 구현하겠습니다!
-- 페이지 주소도 index.html 그대로 두었습니다
+### 멘토님께
+
+- 우선 items 페이지를 home으로 설정해뒀습니다.
+
+- API를 통한 상품 등록은 추후 미션에서 적용한다고 하여 form을 submit할 때 입력폼의 상태값을 콘솔 출력하게 코드를 작성했습니다.
+
 - 아직 리엑트가 익숙하지 않네요😂
 
 ## 스크린샷
 
-|                                               풀스크린                                               |
-| :------------------------------------------------------------------------------------------------: |
-|  ![full](https://github.com/user-attachments/assets/566bc6e6-e2a6-4137-9d0e-b061e226c90f)  |
-
+|                                         풀스크린                                         |
+| :--------------------------------------------------------------------------------------: |
+| ![full](https://github.com/user-attachments/assets/566bc6e6-e2a6-4137-9d0e-b061e226c90f) |
 
 <hr>
 
-|                                              테블릿                                               |
-| :-------------------------------------------------------------------------------------------------: |
+|                                           테블릿                                           |
+| :----------------------------------------------------------------------------------------: |
 | ![tablet](https://github.com/user-attachments/assets/1e76c1ad-4f0f-47d8-8dfc-ef04ddd3d0b3) |
 
 <hr>
 
-|                                              모바일                                               |
-| :-------------------------------------------------------------------------------------------------: |
+|                                           모바일                                           |
+| :----------------------------------------------------------------------------------------: |
 | ![moblie](https://github.com/user-attachments/assets/4515d2ca-37cf-4e61-bab3-b1d9ab9ab888) |
