+++ src/components/ItemForm/ItemForm.css
@@ -1,7 +1,7 @@
 .ItemForm-main {
   display: flex;
   flex-direction: column;
-  padding: 24px 0 69px;
+  padding-bottom: 59px;
 }
 
 .ItemForm-main-title {
@@ -39,4 +39,48 @@
 
 .Tag-main-input {
   margin: 0;
+}
+
+.AddItem-bar {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  margin: 24px 0 24px;
+}
+
+.AddItem-bar-title {
+  font-size: 20px;
+  font-weight: 700;
+  color: #1f2937;
+  margin: 0;
+}
+
+.AddItem-bar-button {
+  width: 74px;
+  height: 42px;
+  padding: 12px 0;
+  border: 0;
+  border-radius: 8px;
+  font-size: 16px;
+  font-weight: 600;
+  color: #f3f4f6;
+  cursor: pointer;
+  background-color: #3692ff;
+}
+
+.AddItem-bar-button:disabled {
+  background-color: #9ca3af;
+  cursor: not-allowed;
+}
+
+@media (max-width: 1200px) {
+  .ItemForm-main {
+    padding-bottom: 78px;
+  }
+}
+
+@media (max-width: 768px) {
+  .ItemForm-main {
+    padding-bottom: 70px;
+  }
 }
