+++ src/components/Tag/Tag.css
@@ -1,5 +1,6 @@
 .Taglist {
   display: flex;
+  flex-wrap: wrap;
   margin-top: 14px;
   margin-bottom: 0px;
   padding: 0;
