+++ src/index.css
@@ -1,0 +1,3 @@
+html {
+  padding: 0;
+}
