+++ src/pages/AddItemPage/AddItemPage.css
@@ -1,35 +1,3 @@
-.AddItem-bar {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  margin: 24px 0 24px;
-}
-
-.AddItem-bar-title {
-  font-size: 20px;
-  font-weight: 700;
-  color: #1f2937;
-  margin: 0;
-}
-
-.AddItem-bar-button {
-  width: 74px;
-  height: 42px;
-  padding: 12px 0;
-  border: 0;
-  border-radius: 8px;
-  font-size: 16px;
-  font-weight: 600;
-  color: #f3f4f6;
-  cursor: pointer;
-  background-color: #3692ff;
-}
-
-.AddItem-bar-button:disabled {
-  background-color: #9ca3af;
-  cursor: not-allowed;
-}
-
 .container {
   width: 1200px;
   margin: 0 auto;
